requirements.txt
# ── Python dependencies ───────────────────────────────────────────────────────
opencv-python>=4.8
mediapipe>=0.10
websocket-client>=1.7
numpy>=1.24
